+++ src/components/layout/QueryLayout/model.ts
@@ -1,6 +1,0 @@
-import type { QueryClientProps } from "@/core/model/api-client";
-
-export type QueryLayoutProps<T> = {
-  queryClientOptions: QueryClientProps;
-  children: (data: T) => React.ReactNode;
-};