+++ src/features/Index.tsx
@@ -1,11 +1,20 @@
-import QueryLayout from "@/components/layout/QueryLayout/QueryLayout";
-import Json from "@/components/utils/Json/Json";
-import type { UsersList } from "@/types/user";
+import QueryLayout from "@components/layout/QueryLayout";
+import Filters from "@components/ui/Filters";
+import { FILTERS_CONFIG } from "@config/filters";
+import { useFiltersResults } from "@hooks/useFilters";
+import type { UsersList, User } from "@type/data/user";
 
 export function Index() {
+    const { setFilters, filteredData } =
+        useFiltersResults<UsersList["users"]>();
+
     return (
         <>
             <h1>Welcome Home!</h1>
+            <Filters
+                filters={FILTERS_CONFIG}
+                onFiltersChange={(values) => setFilters(values)}
+            />
             <QueryLayout<UsersList>
                 queryClientOptions={{
                     queryKeys: ["user"],
@@ -14,7 +23,15 @@ export function Index() {
                     },
                 }}
             >
-                {(data) => <Json obj={data.users} />}
+                {(data) =>
+                    filteredData(data.users).map((user: User) => (
+                        <div key={user.id}>
+                            <p>
+                                {user.firstName} {user.lastName} - {user.role}
+                            </p>
+                        </div>
+                    ))
+                }
             </QueryLayout>
         </>
     );