+++ src/main.tsx
@@ -1,11 +1,10 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
-import "@/index.css";
 import { RouterProvider } from "@tanstack/react-router";
-import { router } from "@/core/router";
 import { QueryClientProvider } from "@tanstack/react-query";
-import { queryClient } from "@/core/query-client";
-
+import { router } from "@utils/router";
+import { queryClient } from "@utils/queryClient";
+import "@styles/index.css";
 createRoot(document.getElementById("root")!).render(
     <StrictMode>
         <QueryClientProvider client={queryClient}>