+++ src/types/core/filters.ts
@@ -1,0 +1,12 @@
+import type { SelectProps, InputTextProps } from "./input";
+
+export type FiltersConfig = {
+  key: string;
+} & ((SelectProps & { type: "select" }) | (InputTextProps & { type: "text" }));
+
+export type FiltersProps = {
+  filters: FiltersConfig[];
+  onFiltersChange: (filters: FilterValues) => void;
+};
+
+export type FilterValues = Record<string, string>;